renamed frontend_logger to logger
