+++ src-tauri/src/logger/cmd.rs
