+++ src-tauri/src/logger/mod.rs
